Added an Anvil, an abstract interactable class, Interactable Manager, Upgradesystem and UpgradeUI
The anvil is created from the Interactable class, the InteractableManager makes the player interact with the closest interactable and shows which is closest. When pressing E near an Anvil a UpgradeUI pops up. The upgradeUI consists of a AbilitySelection window and a UpgradeSelection window. The Upgrade selection windows is opened with the coresponding upgrades to the ability chosen. Both ability's and upgrade's are loaded from scriptable objects.