WormFSM Done
All states a in the WormFSM script. The enemy can dig on impact, which makes it invulnerable. When the player is out of sight it patrols around the map i a random path. The worm has a icon above its head signaling to the player if they are seen or hidden.